Move ModuleRendering to dag package
