+++ engine/src/main/java/org/terasology/rendering/nui/layers/mainMenu/videoSettings/RenderingModuleSettingScreen.java
@@ -21,18 +21,15 @@ import org.terasology.assets.ResourceUrn;
 import org.terasology.context.Context;
 import org.terasology.engine.module.rendering.RenderingModuleManager;
 import org.terasology.i18n.TranslationSystem;
-import org.terasology.math.geom.Vector2i;
 import org.terasology.module.ModuleEnvironment;
-import org.terasology.naming.Name;
 import org.terasology.registry.In;
 import org.terasology.engine.module.rendering.RenderingModuleRegistry;
-import org.terasology.rendering.dag.gsoc.ModuleRendering;
+import org.terasology.rendering.dag.ModuleRendering;
 import org.terasology.rendering.nui.Canvas;
 import org.terasology.rendering.nui.CoreScreenLayer;
 import org.terasology.rendering.nui.WidgetUtil;
 import org.terasology.rendering.nui.databinding.Binding;
 import org.terasology.rendering.nui.databinding.ReadOnlyBinding;
-import org.terasology.rendering.nui.itemRendering.AbstractItemRenderer;
 import org.terasology.rendering.nui.itemRendering.StringTextRenderer;
 import org.terasology.rendering.nui.layers.mainMenu.StartPlayingScreen;
 import org.terasology.rendering.nui.widgets.*;
